Bug fix
After doing some more tests, I found out that block of code will raise
an exception if the library is running for the first time (which means
we have no resources extracted). To deal with it, we simply set
"Extract_Resources" to true and everything will continue they way it
should.